removed bashisms, cleaned script

+++ install.sh
@@ -1,13 +1,18 @@
-#!/usr/bin/env bash
+#!/bin/sh
 # Highly inspired from sontek (John Anderson) dotfiles
 
 if [ ! -d ~/.vimundo ] ; then
     mkdir ~/.vimundo
 fi
 
-function link_file {
-    source="${PWD}/$1"
-    target="${HOME}/${1/_/.}"
+link_file () {
+    source="${1}"
+    target="${2}"
+
+    if [ "$(readlink -f ${target})" = "$(readlink -f ${source})" ] ; then
+        # file already patched, skip
+        return
+    fi
 
     if [ -e "${target}" ]; then
         mv $target $target.bak
@@ -16,22 +21,25 @@ function link_file {
     ln -sf ${source} ${target}
 }
 
+link_home () {
+    source="${PWD}/$1"
+    target="$(echo ${HOME}/${1} | tr '_' '.')"
+    link_file "${source}" "${target}"
+}
+
 if [ "$1" = "vim" ]; then
     for i in _vim*
     do
-       link_file $i
+       link_home $i
     done
 else
     for i in _*
     do
-        link_file $i
+        link_home $i
     done
 fi
 
-if [ -e .kde/share/kde4/services/ServiceMenus/dropboxpublic.desktop ]; then
-    mv .kde/share/kde4/services/ServiceMenus/dropboxpublic.desktop{,.bak}
-fi
-ln -sf ${PWD}/dropboxpublic.desktop ${HOME}.kde/share/kde4/services/ServiceMenus # FIXME: if you need this another time, do a function
+link_file "${PWD}/dropboxpublic.desktop" "${HOME}/.kde/share/kde4/services/ServiceMenus/dropboxpublic.desktop"
 
 git submodule sync
 git submodule init
